move methods which shouldn't be in Item into ItemService
